+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# NEXT
+# 3.1.1
 
 ## New Features